requirements.txt
faster-whisper>=1.0.0
torch>=2.0.0
pyannote.audio>=3.1
numpy
sounddevice
pynput
keyboard
pyperclip
python-dotenv
requests
huggingface-hub
